+++ CHANGELOG.md
@@ -2,13 +2,18 @@
 
 ## [Unreleased][]
 
-[Unreleased]: https://github.com/lueurdev/lueur/compare/0.18.0..HEAD
+[Unreleased]: https://github.com/lueurdev/lueur/compare/0.19.0..HEAD
+
+## [0.19.0][] - 2024-10-15
+
+[0.19.0]: https://github.com/lueurdev/lueur/compare/0.18.0..0.19.0
 
 ### Changed
 
 - Improved links based on communication flow
+- New links from nodes to pods and vice versa
 
-## [0.18.0][] - 2024-10-13
+## [0.18.0][] - 2024-10-14
 
 [0.18.0]: https://github.com/lueurdev/lueur/compare/0.17.1..0.18.0